refactor: rm filter login from eneities mod
