+++ frontend/rust-lib/flowy-grid/src/services/filter/date_filter.rs
@@ -1,0 +1,3 @@
+use crate::entities::GridDateFilter;
+
+impl GridDateFilter {}
