+++ frontend/rust-lib/flowy-grid/src/services/filter/mod.rs
@@ -1,4 +1,9 @@
+mod checkbox_filter;
+mod date_filter;
 mod filter_cache;
 mod filter_service;
+mod number_filter;
+mod select_option_filter;
+mod text_filter;
 
 pub(crate) use filter_service::*;
